requirements.txt
# requirements.txt  ← REPLACES EXISTING FILE
# CHANGE: + python-multipart (required for FastAPI File + Form params)

fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.7.0
httpx>=0.27.0
python-dotenv>=1.0.1
python-multipart>=0.0.9
pytest